+++ library/core/Polygon.cc
@@ -26,6 +26,7 @@
 #include <numeric>
 
 #include <gf/Geometry.h>
+#include <gf/SerializationOps.h>
 #include <gf/Transform.h>
 #include <gf/VectorOps.h>
 
@@ -89,14 +90,16 @@ inline namespace v1 {
   }
 
   bool Polygon::isConvex() const {
-    if (m_points.size() <= 3) {
+    auto size = m_points.size();
+
+    if (size <= 3) {
       return true;
     }
 
     int currentSign = 0;
 
-    for (std::size_t i = 0; i < m_points.size() - 2; ++i) {
-      float x = gf::cross(m_points[i + 1] - m_points[i], m_points[i + 2] - m_points[i + 1]);
+    for (std::size_t i = 0; i < m_points.size(); ++i) {
+      float x = gf::cross(m_points[(i + 1) % size] - m_points[i], m_points[(i + 2) % size] - m_points[(i + 1) % size]);
 
       if (std::abs(x) > Epsilon) {
         int sign = gf::sign(x);
@@ -144,6 +147,34 @@ inline namespace v1 {
 
   void Polygon::simplify(float distance) {
     m_points = gf::simplifyPoints(m_points, distance);
+  }
+
+
+  Serializer& operator|(Serializer& ar, const Polygon& polygon) {
+    uint64_t size = polygon.getPointCount();
+    ar | size;
+
+    for (uint64_t i = 0; i < size; ++i) {
+      Vector2f point = polygon.getPoint(i);
+      ar | point;
+    }
+
+    return ar;
+  }
+
+  Deserializer& operator|(Deserializer& ar, Polygon& polygon) {
+    polygon = Polygon();
+
+    uint64_t size;
+    ar | size;
+
+    for (uint64_t i = 0; i < size; ++i) {
+      Vector2f point;
+      ar | point;
+      polygon.addPoint(point);
+    }
+
+    return ar;
   }
 
 #ifndef DOXYGEN_SHOULD_SKIP_THIS